Conflicts:
	src/process/classes/RawToProfileConverter.java

+++ src/database/subClasses/ExperimentMethods.java
@@ -8,11 +8,17 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
 
-import database.Annotation;
-import database.Experiment;
 import database.FilePathGenerator;
-import database.FileTuple;
+import database.containers.Annotation;
+import database.containers.Experiment;
+import database.containers.FileTuple;
 
+/**
+ * Class that contains all the methods for adding,changing, getting and removing
+ * Experiments in the database. This class is a subClass of databaseAcessor.java
+ *
+ * date: 2014-05-14 version: 1.0
+ */
 public class ExperimentMethods {
 
     private Connection conn;
@@ -29,7 +35,7 @@ public class ExperimentMethods {
 
     /**
      * Gets an experiment from the database.
-     * 
+     *
      * @param expID
      *            the ID of the experiment.
      * @return an Experiment object.
@@ -59,38 +65,43 @@ public class ExperimentMethods {
 
     /**
      * Adds an experiment ID to the database.
-     * 
+     *
      * @param expID
      *            the ID for the experiment.
      * @return the number of tuples inserted in the database.
      * @throws SQLException
      *             if the query does not succeed
-     * @throws IOException 
+     * @throws IOException
      * @throws DuplicatePrimaryKeyException
      *             If the experiment already exists.
      */
     public int addExperiment(String expID) throws SQLException, IOException {
 
-        if (getExperiment(expID) != null) {
-            throw new IOException("This experiment already exists.");
-        }
+    	if (expID == null || expID.isEmpty()) {
+    		throw new IOException("Invalid experiment ID.");
+    	}
+
+    	Experiment e = getExperiment(expID);
+    	if (e != null) {
+    		throw new IOException(expID + " already exists");
+    	}
 
         String query = "INSERT INTO Experiment "
                 + "(ExpID) VALUES (?)";
         PreparedStatement stmt = conn.prepareStatement(query);
         stmt.setString(1, expID);
 
-        fpg.generateExperimentFolders(expID);
-
         int rs = stmt.executeUpdate();
         stmt.close();
+
+        fpg.generateExperimentFolders(expID);
 
         return rs;
     }
 
     /**
      * Deletes an experiment from the database.
-     * 
+     *
      * @param expId
      *            the experiment ID.
      * @param rootDir
@@ -101,9 +112,20 @@ public class ExperimentMethods {
      *             contains at least one file. (All files relating to
      *             an experiment must be deleted first before an
      *             experiment can be deleted from the database)
+     * @throws IOException
      */
-    public int deleteExperiment(String expId, String rootDir)
-            throws SQLException {
+    public int deleteExperiment(String expId)
+            throws SQLException, IOException {
+
+        Experiment e = getExperiment(expId);
+
+        if (e == null) {
+            throw new IOException("No experiment with ID " + expId);
+        }
+
+        if (!e.getFiles().isEmpty()) {
+            throw new IOException("This experiment contains files and therefore cannot be removed.");
+        }
 
         String query = "DELETE FROM Experiment "
                 + "WHERE ExpID ~~* ?";
@@ -114,14 +136,18 @@ public class ExperimentMethods {
         int rs = stmt.executeUpdate();
         stmt.close();
 
-        recursiveDelete(new File(rootDir + expId));
+
+        File experimentFolder = new File(fpg.getRootDirectory() + e.getID());
+        if (experimentFolder.exists()) {
+            recursiveDelete(experimentFolder);
+        }
 
         return rs;
     }
 
     /**
      * Updates a value of a single annotation of a unique experiment
-     * 
+     *
      * @param expID
      *            the name of the experiment to annotate.
      * @param label
@@ -160,7 +186,7 @@ public class ExperimentMethods {
     /**
      * Annotates an experiment with the given label and value. Checks
      * so that the value is valid if it is a drop down annotation.
-     * 
+     *
      * @param expID
      *            the name of the experiment to annotate.
      * @param label
@@ -178,10 +204,6 @@ public class ExperimentMethods {
 
         value = validateAnnotation(label, value);
 
-        if (value == null) {
-            throw new IOException("Invalid annotation");
-        }
-
         String query = "INSERT INTO Annotated_With "
                 + "VALUES (?, ?, ?)";
         PreparedStatement stmt = conn.prepareStatement(query);
@@ -197,7 +219,7 @@ public class ExperimentMethods {
 
     /**
      * Deletes one annotation from a specific experiment.
-     * 
+     *
      * @param expID
      *            the experiment to delete the annotation from.
      * @param label
@@ -225,7 +247,7 @@ public class ExperimentMethods {
     /**
      * Adds all the files that belong to the experiment to an
      * Experiment object.
-     * 
+     *
      * @param e
      *            the experiment to add files to.
      * @return the Experiment object containing all its files.
@@ -251,7 +273,7 @@ public class ExperimentMethods {
     /**
      * Fill an Experiment object with all annotations that exists for
      * that experiment.
-     * 
+     *
      * @param e
      *            the Experiment object.
      * @return the Experiment object containing all it's annotations.
@@ -279,7 +301,7 @@ public class ExperimentMethods {
 
     /**
      * Checks so that the annotation value is valid.
-     * 
+     *
      * @param label
      *            the annotation name.
      * @param value
@@ -287,14 +309,19 @@ public class ExperimentMethods {
      * @return true if the value is valid, else false.
      * @throws SQLException
      *             if the query does not succeed
+     * @throws IOException
      */
     private String validateAnnotation(String label, String value)
-            throws SQLException {
+            throws SQLException, IOException {
 
         Annotation a = annoMethods.getAnnotationObject(label);
 
-        if (a == null || value == null) {
-            return null;
+        if (a == null) {
+            throw new IOException(label + " is not a valid annotation. (Does not exist)");
+        }
+
+        if (value == null) {
+            throw new IOException("Invalid annotation value. (value null)");
         }
 
         if (a.dataType == Annotation.FREETEXT) {
@@ -302,12 +329,17 @@ public class ExperimentMethods {
         }
 
         List<String> choices = a.getPossibleValues();
-        value = matchChoice(value, choices);
+        String res = matchChoice(value, choices);
 
-        return value;
+        if (res == null) {
+            throw new IOException(value + "is not a valid choice for the drop down annotation " + label);
+        }
+
+        return res;
     }
 
     private String matchChoice(String value, List<String> choices) {
+
         for (String s : choices) {
             if (value.equalsIgnoreCase(s)) {
                 return s;
@@ -319,12 +351,14 @@ public class ExperimentMethods {
     /**
      * Recursively deletes a folder with all it's subfolders and
      * files.
-     * 
+     *
      * @param folder
      *            the folder to delete.
      */
     private static void recursiveDelete(File folder) {
+
         File[] contents = folder.listFiles();
+
         if (contents == null || contents.length == 0) {
             folder.delete();
         } else {
@@ -332,7 +366,7 @@ public class ExperimentMethods {
                 recursiveDelete(f);
             }
         }
+
         folder.delete();
     }
-
 }
